+++ extensions/omni_kube/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2025-09-14
+
+### Added
+
+* Full resource schema population (view or tables) [#949](https://github.com/omnigres/omnigres/pull/949)
+* `omni_kube` role to simplify granting the capabilities to users [#951](https://github.com/omnigres/omnigres/pull/951)
+
+### Fixed
+
+* Restore search path upon resource table and view creation [#949](https://github.com/omnigres/omnigres/pull/949)
+
 ## [0.3.0] - 2025-09-03
 
 ### Added
@@ -58,3 +69,5 @@ Initial release
 [0.2.0]: [https://github.com/omnigres/omnigres/pull/913]
 
 [0.3.0]: [https://github.com/omnigres/omnigres/pull/922]
+
+[0.4.0]: [https://github.com/omnigres/omnigres/pull/948]